Move External Volumes reducer
